Move namespaces to their own modules
