+++ py-polars/polars/internals/expr/expr.py
@@ -1,24 +1,18 @@
 from __future__ import annotations
 
-import copy
 import math
 import random
-from datetime import date, datetime, timedelta
+from datetime import date, datetime
 from typing import TYPE_CHECKING, Any, Callable, Sequence
 
 from polars import internals as pli
-from polars.datatypes import (
-    DTYPE_TEMPORAL_UNITS,
-    DataType,
-    Date,
-    Datetime,
-    Float64,
-    Int32,
-    Time,
-    UInt32,
-    py_type_to_dtype,
-)
-from polars.utils import _timedelta_to_pl_duration, is_expr_sequence, is_pyexpr_sequence
+from polars.datatypes import DataType, Datetime, Float64, UInt32, py_type_to_dtype
+from polars.internals.expr.categorical import ExprCatNameSpace
+from polars.internals.expr.datetime import ExprDateTimeNameSpace
+from polars.internals.expr.list import ExprListNameSpace
+from polars.internals.expr.string import ExprStringNameSpace
+from polars.internals.expr.struct import ExprStructNameSpace
+from polars.utils import is_expr_sequence, is_pyexpr_sequence
 
 try:
     from polars.polars import PyExpr
@@ -36,16 +30,11 @@ except ImportError:
 
 if TYPE_CHECKING:
     from polars.internals.type_aliases import (
-        CategoricalOrdering,
         ClosedWindow,
-        EpochTimeUnit,
         FillNullStrategy,
         InterpolationMethod,
         NullBehavior,
         RankMethod,
-        TimeUnit,
-        ToStructStrategy,
-        TransferEncoding,
     )
 
 
@@ -5088,1994 +5077,6 @@ class Expr:
         return ExprStructNameSpace(self)
 
 
-class ExprStructNameSpace:
-    """Namespace for struct related expressions."""
-
-    def __init__(self, expr: Expr):
-        self._pyexpr = expr._pyexpr
-
-    def field(self, name: str) -> Expr:
-        """
-        Retrieve one of the fields of this `Struct` as a new Series.
-
-        Parameters
-        ----------
-        name
-            Name of the field
-
-        Examples
-        --------
-        >>> df = (
-        ...     pl.DataFrame(
-        ...         {
-        ...             "int": [1, 2],
-        ...             "str": ["a", "b"],
-        ...             "bool": [True, None],
-        ...             "list": [[1, 2], [3]],
-        ...         }
-        ...     )
-        ...     .to_struct("my_struct")
-        ...     .to_frame()
-        ... )
-        >>> df.select(pl.col("my_struct").struct.field("str"))
-        shape: (2, 1)
-        ┌─────┐
-        │ str │
-        │ --- │
-        │ str │
-        ╞═════╡
-        │ a   │
-        ├╌╌╌╌╌┤
-        │ b   │
-        └─────┘
-
-        """
-        return wrap_expr(self._pyexpr.struct_field_by_name(name))
-
-    def rename_fields(self, names: list[str]) -> Expr:
-        """
-        Rename the fields of the struct
-
-        Parameters
-        ----------
-        names
-            New names in the order of the struct's fields
-
-        Examples
-        --------
-        >>> df = (
-        ...     pl.DataFrame(
-        ...         {
-        ...             "int": [1, 2],
-        ...             "str": ["a", "b"],
-        ...             "bool": [True, None],
-        ...             "list": [[1, 2], [3]],
-        ...         }
-        ...     )
-        ...     .to_struct("my_struct")
-        ...     .to_frame()
-        ... )
-        >>> df = df.with_column(
-        ...     pl.col("my_struct").struct.rename_fields(["INT", "STR", "BOOL", "LIST"])
-        ... )
-
-        Does NOT work anymore:
-        # df.select(pl.col("my_struct").struct.field("int"))
-        #               PanicException: int not found ^^^
-
-        >>> df.select(pl.col("my_struct").struct.field("INT"))
-        shape: (2, 1)
-        ┌─────┐
-        │ INT │
-        │ --- │
-        │ i64 │
-        ╞═════╡
-        │ 1   │
-        ├╌╌╌╌╌┤
-        │ 2   │
-        └─────┘
-
-        """
-        return wrap_expr(self._pyexpr.struct_rename_fields(names))
-
-
-class ExprListNameSpace:
-    """Namespace for list related expressions."""
-
-    def __init__(self, expr: Expr):
-        self._pyexpr = expr._pyexpr
-
-    def lengths(self) -> Expr:
-        """
-        Get the length of the arrays as UInt32.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": [1, 2], "bar": [["a", "b"], ["c"]]})
-        >>> df.select(pl.col("bar").arr.lengths())
-        shape: (2, 1)
-        ┌─────┐
-        │ bar │
-        │ --- │
-        │ u32 │
-        ╞═════╡
-        │ 2   │
-        ├╌╌╌╌╌┤
-        │ 1   │
-        └─────┘
-
-        """
-        return wrap_expr(self._pyexpr.arr_lengths())
-
-    def sum(self) -> Expr:
-        """Sum all the arrays in the list."""
-        return wrap_expr(self._pyexpr.lst_sum())
-
-    def max(self) -> Expr:
-        """Compute the max value of the arrays in the list."""
-        return wrap_expr(self._pyexpr.lst_max())
-
-    def min(self) -> Expr:
-        """Compute the min value of the arrays in the list."""
-        return wrap_expr(self._pyexpr.lst_min())
-
-    def mean(self) -> Expr:
-        """Compute the mean value of the arrays in the list."""
-        return wrap_expr(self._pyexpr.lst_mean())
-
-    def sort(self, reverse: bool = False) -> Expr:
-        """
-        Sort the arrays in the list
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {
-        ...         "a": [[3, 2, 1], [9, 1, 2]],
-        ...     }
-        ... )
-        >>> df.select(pl.col("a").arr.sort())
-        shape: (2, 1)
-        ┌───────────┐
-        │ a         │
-        │ ---       │
-        │ list[i64] │
-        ╞═══════════╡
-        │ [1, 2, 3] │
-        ├╌╌╌╌╌╌╌╌╌╌╌┤
-        │ [1, 2, 9] │
-        └───────────┘
-
-        """
-        return wrap_expr(self._pyexpr.lst_sort(reverse))
-
-    def reverse(self) -> Expr:
-        """
-        Reverse the arrays in the list
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {
-        ...         "a": [[3, 2, 1], [9, 1, 2]],
-        ...     }
-        ... )
-        >>> df.select(pl.col("a").arr.reverse())
-        shape: (2, 1)
-        ┌───────────┐
-        │ a         │
-        │ ---       │
-        │ list[i64] │
-        ╞═══════════╡
-        │ [1, 2, 3] │
-        ├╌╌╌╌╌╌╌╌╌╌╌┤
-        │ [2, 1, 9] │
-        └───────────┘
-
-        """
-        return wrap_expr(self._pyexpr.lst_reverse())
-
-    def unique(self) -> Expr:
-        """Get the unique/distinct values in the list."""
-        return wrap_expr(self._pyexpr.lst_unique())
-
-    def concat(
-        self, other: list[Expr | str] | Expr | str | pli.Series | list[Any]
-    ) -> Expr:
-        """
-        Concat the arrays in a Series dtype List in linear time.
-
-        Parameters
-        ----------
-        other
-            Columns to concat into a List Series
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {
-        ...         "a": [["a"], ["x"]],
-        ...         "b": [["b", "c"], ["y", "z"]],
-        ...     }
-        ... )
-        >>> df.select(pl.col("a").arr.concat("b"))
-        shape: (2, 1)
-        ┌─────────────────┐
-        │ a               │
-        │ ---             │
-        │ list[str]       │
-        ╞═════════════════╡
-        │ ["a", "b", "c"] │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ ["x", "y", "z"] │
-        └─────────────────┘
-
-        """
-        if isinstance(other, list) and (
-            not isinstance(other[0], (Expr, str, pli.Series))
-        ):
-            return self.concat(pli.Series([other]))
-
-        other_list: list[Expr | str | pli.Series]
-        if not isinstance(other, list):
-            other_list = [other]
-        else:
-            other_list = copy.copy(other)  # type: ignore[arg-type]
-
-        other_list.insert(0, wrap_expr(self._pyexpr))
-        return pli.concat_list(other_list)
-
-    def get(self, index: int) -> Expr:
-        """
-        Get the value by index in the sublists.
-        So index `0` would return the first item of every sublist
-        and index `-1` would return the last item of every sublist
-        if an index is out of bounds, it will return a `None`.
-
-        Parameters
-        ----------
-        index
-            Index to return per sublist
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": [[3, 2, 1], [], [1, 2]]})
-        >>> df.select(pl.col("foo").arr.get(0))
-        shape: (3, 1)
-        ┌──────┐
-        │ foo  │
-        │ ---  │
-        │ i64  │
-        ╞══════╡
-        │ 3    │
-        ├╌╌╌╌╌╌┤
-        │ null │
-        ├╌╌╌╌╌╌┤
-        │ 1    │
-        └──────┘
-
-        """
-        return wrap_expr(self._pyexpr.lst_get(index))
-
-    def first(self) -> Expr:
-        """
-        Get the first value of the sublists.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": [[3, 2, 1], [], [1, 2]]})
-        >>> df.select(pl.col("foo").arr.first())
-        shape: (3, 1)
-        ┌──────┐
-        │ foo  │
-        │ ---  │
-        │ i64  │
-        ╞══════╡
-        │ 3    │
-        ├╌╌╌╌╌╌┤
-        │ null │
-        ├╌╌╌╌╌╌┤
-        │ 1    │
-        └──────┘
-
-        """
-        return self.get(0)
-
-    def last(self) -> Expr:
-        """
-        Get the last value of the sublists.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": [[3, 2, 1], [], [1, 2]]})
-        >>> df.select(pl.col("foo").arr.last())
-        shape: (3, 1)
-        ┌──────┐
-        │ foo  │
-        │ ---  │
-        │ i64  │
-        ╞══════╡
-        │ 1    │
-        ├╌╌╌╌╌╌┤
-        │ null │
-        ├╌╌╌╌╌╌┤
-        │ 2    │
-        └──────┘
-
-        """
-        return self.get(-1)
-
-    def contains(self, item: float | str | bool | int | date | datetime | Expr) -> Expr:
-        """
-        Check if sublists contain the given item.
-
-        Parameters
-        ----------
-        item
-            Item that will be checked for membership
-
-        Returns
-        -------
-        Boolean mask
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": [[3, 2, 1], [], [1, 2]]})
-        >>> df.select(pl.col("foo").arr.contains(1))
-        shape: (3, 1)
-        ┌───────┐
-        │ foo   │
-        │ ---   │
-        │ bool  │
-        ╞═══════╡
-        │ true  │
-        ├╌╌╌╌╌╌╌┤
-        │ false │
-        ├╌╌╌╌╌╌╌┤
-        │ true  │
-        └───────┘
-
-        """
-        return wrap_expr(self._pyexpr.arr_contains(expr_to_lit_or_expr(item)._pyexpr))
-
-    def join(self, separator: str) -> Expr:
-        """
-        Join all string items in a sublist and place a separator between them.
-        This errors if inner type of list `!= Utf8`.
-
-        Parameters
-        ----------
-        separator
-            string to separate the items with
-
-        Returns
-        -------
-        Series of dtype Utf8
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"s": [["a", "b", "c"], ["x", "y"]]})
-        >>> df.select(pl.col("s").arr.join(" "))
-        shape: (2, 1)
-        ┌───────┐
-        │ s     │
-        │ ---   │
-        │ str   │
-        ╞═══════╡
-        │ a b c │
-        ├╌╌╌╌╌╌╌┤
-        │ x y   │
-        └───────┘
-
-        """
-        return wrap_expr(self._pyexpr.lst_join(separator))
-
-    def arg_min(self) -> Expr:
-        """
-        Retrieve the index of the minimal value in every sublist
-
-        Returns
-        -------
-        Series of dtype UInt32/UInt64 (depending on compilation)
-
-        """
-        return wrap_expr(self._pyexpr.lst_arg_min())
-
-    def arg_max(self) -> Expr:
-        """
-        Retrieve the index of the maximum value in every sublist
-
-        Returns
-        -------
-        Series of dtype UInt32/UInt64 (depending on compilation)
-
-        """
-        return wrap_expr(self._pyexpr.lst_arg_max())
-
-    def diff(self, n: int = 1, null_behavior: NullBehavior = "ignore") -> Expr:
-        """
-        Calculate the n-th discrete difference of every sublist.
-
-        Parameters
-        ----------
-        n
-            Number of slots to shift.
-        null_behavior : {'ignore', 'drop'}
-            How to handle null values.
-
-        Examples
-        --------
-        >>> s = pl.Series("a", [[1, 2, 3, 4], [10, 2, 1]])
-        >>> s.arr.diff()
-        shape: (2,)
-        Series: 'a' [list]
-        [
-            [null, 1, ... 1]
-            [null, -8, -1]
-        ]
-
-        """
-        return wrap_expr(self._pyexpr.lst_diff(n, null_behavior))
-
-    def shift(self, periods: int = 1) -> Expr:
-        """
-        Shift the values by a given period and fill the parts that will be empty due to
-        this operation with nulls.
-
-        Parameters
-        ----------
-        periods
-            Number of places to shift (may be negative).
-
-        Examples
-        --------
-        >>> s = pl.Series("a", [[1, 2, 3, 4], [10, 2, 1]])
-        >>> s.arr.shift()
-        shape: (2,)
-        Series: 'a' [list]
-        [
-            [null, 1, ... 3]
-            [null, 10, 2]
-        ]
-
-        """
-        return wrap_expr(self._pyexpr.lst_shift(periods))
-
-    def slice(self, offset: int, length: int) -> Expr:
-        """
-        Slice every sublist
-
-        Parameters
-        ----------
-        offset
-            Take the values from this index offset.
-        length
-            The length of the slice to take.
-
-        Examples
-        --------
-        >>> s = pl.Series("a", [[1, 2, 3, 4], [10, 2, 1]])
-        >>> s.arr.slice(1, 2)
-        shape: (2,)
-        Series: 'a' [list]
-        [
-            [2, 3]
-            [2, 1]
-        ]
-
-        """
-        return wrap_expr(self._pyexpr.lst_slice(offset, length))
-
-    def head(self, n: int = 5) -> Expr:
-        """
-        Slice the head of every sublist
-
-        Parameters
-        ----------
-        n
-            How many values to take in the slice.
-
-        Examples
-        --------
-        >>> s = pl.Series("a", [[1, 2, 3, 4], [10, 2, 1]])
-        >>> s.arr.head(2)
-        shape: (2,)
-        Series: 'a' [list]
-        [
-            [1, 2]
-            [10, 2]
-        ]
-
-        """
-        return self.slice(0, n)
-
-    def tail(self, n: int = 5) -> Expr:
-        """
-        Slice the tail of every sublist
-
-        Parameters
-        ----------
-        n
-            How many values to take in the slice.
-
-        Examples
-        --------
-        >>> s = pl.Series("a", [[1, 2, 3, 4], [10, 2, 1]])
-        >>> s.arr.tail(2)
-        shape: (2,)
-        Series: 'a' [list]
-        [
-            [3, 4]
-            [2, 1]
-        ]
-
-        """
-        return self.slice(-n, n)
-
-    def to_struct(
-        self,
-        n_field_strategy: ToStructStrategy = "first_non_null",
-        name_generator: Callable[[int], str] | None = None,
-    ) -> Expr:
-        """
-        Convert the series of type ``List`` to a series of type ``Struct``.
-
-        Parameters
-        ----------
-        n_field_strategy : {'first_non_null', 'max_width'}
-            Strategy to determine the number of fields of the struct.
-        name_generator
-            A custom function that can be used to generate the field names.
-            Default field names are `field_0, field_1 .. field_n`
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"a": [[1, 2, 3], [1, 2]]})
-        >>> df.select([pl.col("a").arr.to_struct()])
-        shape: (2, 1)
-        ┌────────────┐
-        │ a          │
-        │ ---        │
-        │ struct[3]  │
-        ╞════════════╡
-        │ {1,2,3}    │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ {1,2,null} │
-        └────────────┘
-        >>> df.select(
-        ...     [
-        ...         pl.col("a").arr.to_struct(
-        ...             name_generator=lambda idx: f"col_name_{idx}"
-        ...         )
-        ...     ]
-        ... ).to_series().to_list()
-        [{'col_name_0': 1, 'col_name_1': 2, 'col_name_2': 3},
-        {'col_name_0': 1, 'col_name_1': 2, 'col_name_2': None}]
-
-        """
-        return wrap_expr(self._pyexpr.lst_to_struct(n_field_strategy, name_generator))
-
-    def eval(self, expr: Expr, parallel: bool = False) -> Expr:
-        """
-        Run any polars expression against the lists' elements
-
-        Parameters
-        ----------
-        expr
-            Expression to run. Note that you can select an element with `pl.first()`, or
-            `pl.col()`
-        parallel
-            Run all expression parallel. Don't activate this blindly.
-            Parallelism is worth it if there is enough work to do per thread.
-
-            This likely should not be use in the groupby context, because we already
-            parallel execution per group
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"a": [1, 8, 3], "b": [4, 5, 2]})
-        >>> df.with_column(
-        ...     pl.concat_list(["a", "b"]).arr.eval(pl.element().rank()).alias("rank")
-        ... )
-        shape: (3, 3)
-        ┌─────┬─────┬────────────┐
-        │ a   ┆ b   ┆ rank       │
-        │ --- ┆ --- ┆ ---        │
-        │ i64 ┆ i64 ┆ list[f32]  │
-        ╞═════╪═════╪════════════╡
-        │ 1   ┆ 4   ┆ [1.0, 2.0] │
-        ├╌╌╌╌╌┼╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ 8   ┆ 5   ┆ [2.0, 1.0] │
-        ├╌╌╌╌╌┼╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ 3   ┆ 2   ┆ [2.0, 1.0] │
-        └─────┴─────┴────────────┘
-
-        """
-        return wrap_expr(self._pyexpr.lst_eval(expr._pyexpr, parallel))
-
-
-class ExprStringNameSpace:
-    """Namespace for string related expressions."""
-
-    def __init__(self, expr: Expr):
-        self._pyexpr = expr._pyexpr
-
-    def strptime(
-        self,
-        datatype: type[Date] | type[Datetime] | type[Time],
-        fmt: str | None = None,
-        strict: bool = True,
-        exact: bool = True,
-    ) -> Expr:
-        """
-        Parse a Utf8 expression to a Date/Datetime/Time type.
-
-        Parameters
-        ----------
-        datatype
-            Date | Datetime | Time.
-        fmt
-            Format to use, refer to the `chrono strftime documentation
-            <https://docs.rs/chrono/latest/chrono/format/strftime/index.html>`_
-            for specification. Example: ``"%y-%m-%d"``.
-        strict
-            Raise an error if any conversion fails.
-        exact
-            - If True, require an exact format match.
-            - If False, allow the format to match anywhere in the target string.
-
-        Examples
-        --------
-        Dealing with different formats.
-
-        >>> s = pl.Series(
-        ...     "date",
-        ...     [
-        ...         "2021-04-22",
-        ...         "2022-01-04 00:00:00",
-        ...         "01/31/22",
-        ...         "Sun Jul  8 00:34:60 2001",
-        ...     ],
-        ... )
-        >>> (
-        ...     s.to_frame().with_column(
-        ...         pl.col("date")
-        ...         .str.strptime(pl.Date, "%F", strict=False)
-        ...         .fill_null(
-        ...             pl.col("date").str.strptime(pl.Date, "%F %T", strict=False)
-        ...         )
-        ...         .fill_null(pl.col("date").str.strptime(pl.Date, "%D", strict=False))
-        ...         .fill_null(pl.col("date").str.strptime(pl.Date, "%c", strict=False))
-        ...     )
-        ... )
-        shape: (4, 1)
-        ┌────────────┐
-        │ date       │
-        │ ---        │
-        │ date       │
-        ╞════════════╡
-        │ 2021-04-22 │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ 2022-01-04 │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ 2022-01-31 │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ 2001-07-08 │
-        └────────────┘
-
-        """
-        if not issubclass(datatype, DataType):  # pragma: no cover
-            raise ValueError(f"expected: {DataType} got: {datatype}")
-        if datatype == Date:
-            return wrap_expr(self._pyexpr.str_parse_date(fmt, strict, exact))
-        elif datatype == Datetime:
-            return wrap_expr(self._pyexpr.str_parse_datetime(fmt, strict, exact))
-        elif datatype == Time:
-            return wrap_expr(self._pyexpr.str_parse_time(fmt, strict, exact))
-        else:  # pragma: no cover
-            raise ValueError("dtype should be of type {Date, Datetime, Time}")
-
-    def lengths(self) -> Expr:
-        """
-        Get the length of the Strings as UInt32.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"s": [None, "bears", "110"]})
-        >>> df.select(["s", pl.col("s").str.lengths().alias("len")])
-        shape: (3, 2)
-        ┌───────┬──────┐
-        │ s     ┆ len  │
-        │ ---   ┆ ---  │
-        │ str   ┆ u32  │
-        ╞═══════╪══════╡
-        │ null  ┆ null │
-        ├╌╌╌╌╌╌╌┼╌╌╌╌╌╌┤
-        │ bears ┆ 5    │
-        ├╌╌╌╌╌╌╌┼╌╌╌╌╌╌┤
-        │ 110   ┆ 3    │
-        └───────┴──────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_lengths())
-
-    def concat(self, delimiter: str = "-") -> Expr:
-        """
-        Vertically concat the values in the Series to a single string value.
-
-        Parameters
-        ----------
-        delimiter
-            The delimiter to insert between consecutive string values.
-
-        Returns
-        -------
-        Series of dtype Utf8
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": [1, None, 2]})
-        >>> df.select(pl.col("foo").str.concat("-"))
-        shape: (1, 1)
-        ┌──────────┐
-        │ foo      │
-        │ ---      │
-        │ str      │
-        ╞══════════╡
-        │ 1-null-2 │
-        └──────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_concat(delimiter))
-
-    def to_uppercase(self) -> Expr:
-        """Transform to uppercase variant."""
-        return wrap_expr(self._pyexpr.str_to_uppercase())
-
-    def to_lowercase(self) -> Expr:
-        """Transform to lowercase variant."""
-        return wrap_expr(self._pyexpr.str_to_lowercase())
-
-    def strip(self) -> Expr:
-        """Remove leading and trailing whitespace."""
-        return wrap_expr(self._pyexpr.str_strip())
-
-    def lstrip(self) -> Expr:
-        """Remove leading whitespace."""
-        return wrap_expr(self._pyexpr.str_lstrip())
-
-    def rstrip(self) -> Expr:
-        """Remove trailing whitespace."""
-        return wrap_expr(self._pyexpr.str_rstrip())
-
-    def zfill(self, alignment: int) -> Expr:
-        """
-        Return a copy of the string left filled with ASCII '0' digits to make a string
-        of length width.
-
-        A leading sign prefix ('+'/'-') is handled by inserting the padding after the
-        sign character rather than before. The original string is returned if width is
-        less than or equal to ``len(s)``.
-
-        Parameters
-        ----------
-        alignment
-            Fill the value up to this length
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {
-        ...         "num": [-10, -1, 0, 1, 10, 100, 1000, 10000, 100000, 1000000, None],
-        ...     }
-        ... )
-        >>> df.with_column(pl.col("num").cast(str).str.zfill(5))
-        shape: (11, 1)
-        ┌─────────┐
-        │ num     │
-        │ ---     │
-        │ str     │
-        ╞═════════╡
-        │ -0010   │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ -0001   │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ 00000   │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ 00001   │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ ...     │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ 10000   │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ 100000  │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ 1000000 │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ null    │
-        └─────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_zfill(alignment))
-
-    def ljust(self, width: int, fillchar: str = " ") -> Expr:
-        """
-        Return the string left justified in a string of length ``width``.
-
-        Padding is done using the specified ``fillchar``.
-        The original string is returned if ``width`` is less than or equal to
-        ``len(s)``.
-
-        Parameters
-        ----------
-        width
-            Justify left to this length.
-        fillchar
-            Fill with this ASCII character.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"a": ["cow", "monkey", None, "hippopotamus"]})
-        >>> df.select(pl.col("a").str.ljust(8, "*"))
-        shape: (4, 1)
-        ┌──────────────┐
-        │ a            │
-        │ ---          │
-        │ str          │
-        ╞══════════════╡
-        │ cow*****     │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ monkey**     │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ null         │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ hippopotamus │
-        └──────────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_ljust(width, fillchar))
-
-    def rjust(self, width: int, fillchar: str = " ") -> Expr:
-        """
-        Return the string right justified in a string of length ``width``.
-
-        Padding is done using the specified ``fillchar``.
-        The original string is returned if ``width`` is less than or equal to
-        ``len(s)``.
-
-        Parameters
-        ----------
-        width
-            Justify right to this length.
-        fillchar
-            Fill with this ASCII character.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"a": ["cow", "monkey", None, "hippopotamus"]})
-        >>> df.select(pl.col("a").str.rjust(8, "*"))
-        shape: (4, 1)
-        ┌──────────────┐
-        │ a            │
-        │ ---          │
-        │ str          │
-        ╞══════════════╡
-        │ *****cow     │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ **monkey     │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ null         │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ hippopotamus │
-        └──────────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_rjust(width, fillchar))
-
-    def contains(self, pattern: str, literal: bool = False) -> Expr:
-        """
-        Check if string contains a substring that matches a regex.
-
-        Parameters
-        ----------
-        pattern
-            A valid regex pattern.
-        literal
-            Treat pattern as a literal string.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"a": ["Crab", "cat and dog", "rab$bit", None]})
-        >>> df.select(
-        ...     [
-        ...         pl.col("a"),
-        ...         pl.col("a").str.contains("cat|bit").alias("regex"),
-        ...         pl.col("a").str.contains("rab$", literal=True).alias("literal"),
-        ...     ]
-        ... )
-        shape: (4, 3)
-        ┌─────────────┬───────┬─────────┐
-        │ a           ┆ regex ┆ literal │
-        │ ---         ┆ ---   ┆ ---     │
-        │ str         ┆ bool  ┆ bool    │
-        ╞═════════════╪═══════╪═════════╡
-        │ Crab        ┆ false ┆ false   │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┤
-        │ cat and dog ┆ true  ┆ false   │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┤
-        │ rab$bit     ┆ true  ┆ true    │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┤
-        │ null        ┆ null  ┆ null    │
-        └─────────────┴───────┴─────────┘
-
-        See Also
-        --------
-        starts_with : Check if string values start with a substring.
-        ends_with : Check if string values end with a substring.
-
-        """
-        return wrap_expr(self._pyexpr.str_contains(pattern, literal))
-
-    def ends_with(self, sub: str) -> Expr:
-        """
-        Check if string values end with a substring.
-
-        Parameters
-        ----------
-        sub
-            Suffix substring.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"fruits": ["apple", "mango", None]})
-        >>> df.with_column(
-        ...     pl.col("fruits").str.ends_with("go").alias("has_suffix"),
-        ... )
-        shape: (3, 2)
-        ┌────────┬────────────┐
-        │ fruits ┆ has_suffix │
-        │ ---    ┆ ---        │
-        │ str    ┆ bool       │
-        ╞════════╪════════════╡
-        │ apple  ┆ false      │
-        ├╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ mango  ┆ true       │
-        ├╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ null   ┆ null       │
-        └────────┴────────────┘
-
-        Using ``ends_with`` as a filter condition:
-
-        >>> df.filter(pl.col("fruits").str.ends_with("go"))
-        shape: (1, 1)
-        ┌────────┐
-        │ fruits │
-        │ ---    │
-        │ str    │
-        ╞════════╡
-        │ mango  │
-        └────────┘
-
-        See Also
-        --------
-        contains : Check if string contains a substring that matches a regex.
-        starts_with : Check if string values start with a substring.
-
-        """
-        return wrap_expr(self._pyexpr.str_ends_with(sub))
-
-    def starts_with(self, sub: str) -> Expr:
-        """
-        Check if string values start with a substring.
-
-        Parameters
-        ----------
-        sub
-            Prefix substring.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"fruits": ["apple", "mango", None]})
-        >>> df.with_column(
-        ...     pl.col("fruits").str.starts_with("app").alias("has_prefix"),
-        ... )
-        shape: (3, 2)
-        ┌────────┬────────────┐
-        │ fruits ┆ has_prefix │
-        │ ---    ┆ ---        │
-        │ str    ┆ bool       │
-        ╞════════╪════════════╡
-        │ apple  ┆ true       │
-        ├╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ mango  ┆ false      │
-        ├╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ null   ┆ null       │
-        └────────┴────────────┘
-
-        Using ``starts_with`` as a filter condition:
-
-        >>> df.filter(pl.col("fruits").str.starts_with("app"))
-        shape: (1, 1)
-        ┌────────┐
-        │ fruits │
-        │ ---    │
-        │ str    │
-        ╞════════╡
-        │ apple  │
-        └────────┘
-
-        See Also
-        --------
-        contains : Check if string contains a substring that matches a regex.
-        ends_with : Check if string values end with a substring.
-
-        """
-        return wrap_expr(self._pyexpr.str_starts_with(sub))
-
-    def json_path_match(self, json_path: str) -> Expr:
-        """
-        Extract the first match of json string with provided JSONPath expression.
-        Throw errors if encounter invalid json strings.
-        All return value will be casted to Utf8 regardless of the original value.
-
-        Documentation on JSONPath standard can be found
-        `here <https://goessner.net/articles/JsonPath/>`_.
-
-        Parameters
-        ----------
-        json_path
-            A valid JSON path query string.
-
-        Returns
-        -------
-        Utf8 array. Contain null if original value is null or the json_path return
-        nothing.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {"json_val": ['{"a":"1"}', None, '{"a":2}', '{"a":2.1}', '{"a":true}']}
-        ... )
-        >>> df.select(pl.col("json_val").str.json_path_match("$.a"))
-        shape: (5, 1)
-        ┌──────────┐
-        │ json_val │
-        │ ---      │
-        │ str      │
-        ╞══════════╡
-        │ 1        │
-        ├╌╌╌╌╌╌╌╌╌╌┤
-        │ null     │
-        ├╌╌╌╌╌╌╌╌╌╌┤
-        │ 2        │
-        ├╌╌╌╌╌╌╌╌╌╌┤
-        │ 2.1      │
-        ├╌╌╌╌╌╌╌╌╌╌┤
-        │ true     │
-        └──────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_json_path_match(json_path))
-
-    def decode(self, encoding: TransferEncoding, strict: bool = False) -> Expr:
-        """
-        Decode a value using the provided encoding.
-
-        Parameters
-        ----------
-        encoding : {'hex', 'base64'}
-            The encoding to use.
-        strict
-            How to handle invalid inputs:
-
-            - ``True``: An error will be thrown if unable to decode a value.
-            - ``False``: Unhandled values will be replaced with `None`.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"encoded": ["666f6f", "626172", None]})
-        >>> df.select(pl.col("encoded").str.decode("hex"))
-        shape: (3, 1)
-        ┌─────────┐
-        │ encoded │
-        │ ---     │
-        │ str     │
-        ╞═════════╡
-        │ foo     │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ bar     │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ null    │
-        └─────────┘
-
-        """
-        if encoding == "hex":
-            return wrap_expr(self._pyexpr.str_hex_decode(strict))
-        elif encoding == "base64":
-            return wrap_expr(self._pyexpr.str_base64_decode(strict))
-        else:
-            raise ValueError(
-                f"encoding must be one of {{'hex', 'base64'}}, got {encoding}"
-            )
-
-    def encode(self, encoding: TransferEncoding) -> Expr:
-        """
-        Encode a value using the provided encoding.
-
-        Parameters
-        ----------
-        encoding : {'hex', 'base64'}
-            The encoding to use.
-
-        Returns
-        -------
-        Utf8 array with values encoded using provided encoding
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"strings": ["foo", "bar", None]})
-        >>> df.select(pl.col("strings").str.encode("hex"))
-        shape: (3, 1)
-        ┌─────────┐
-        │ strings │
-        │ ---     │
-        │ str     │
-        ╞═════════╡
-        │ 666f6f  │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ 626172  │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ null    │
-        └─────────┘
-
-        """
-        if encoding == "hex":
-            return wrap_expr(self._pyexpr.str_hex_encode())
-        elif encoding == "base64":
-            return wrap_expr(self._pyexpr.str_base64_encode())
-        else:
-            raise ValueError(
-                f"encoding must be one of {{'hex', 'base64'}}, got {encoding}"
-            )
-
-    def extract(self, pattern: str, group_index: int = 1) -> Expr:
-        r"""
-        Extract the target capture group from provided patterns.
-
-        Parameters
-        ----------
-        pattern
-            A valid regex pattern
-        group_index
-            Index of the targeted capture group.
-            Group 0 mean the whole pattern, first group begin at index 1
-            Default to the first capture group
-
-        Returns
-        -------
-        Utf8 array. Contain null if original value is null or regex capture nothing.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {
-        ...         "a": [
-        ...             "http://vote.com/ballon_dor?candidate=messi&ref=polars",
-        ...             "http://vote.com/ballon_dor?candidat=jorginho&ref=polars",
-        ...             "http://vote.com/ballon_dor?candidate=ronaldo&ref=polars",
-        ...         ]
-        ...     }
-        ... )
-        >>> df.select(
-        ...     [
-        ...         pl.col("a").str.extract(r"candidate=(\w+)", 1),
-        ...     ]
-        ... )
-        shape: (3, 1)
-        ┌─────────┐
-        │ a       │
-        │ ---     │
-        │ str     │
-        ╞═════════╡
-        │ messi   │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ null    │
-        ├╌╌╌╌╌╌╌╌╌┤
-        │ ronaldo │
-        └─────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_extract(pattern, group_index))
-
-    def extract_all(self, pattern: str) -> Expr:
-        r"""
-        Extract each successive non-overlapping regex match in an individual string as
-        an array.
-
-        Parameters
-        ----------
-        pattern
-            A valid regex pattern
-
-        Returns
-        -------
-        List[Utf8] array. Contain null if original value is null or regex capture
-        nothing.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": ["123 bla 45 asd", "xyz 678 910t"]})
-        >>> df.select(
-        ...     [
-        ...         pl.col("foo").str.extract_all(r"(\d+)").alias("extracted_nrs"),
-        ...     ]
-        ... )
-        shape: (2, 1)
-        ┌────────────────┐
-        │ extracted_nrs  │
-        │ ---            │
-        │ list[str]      │
-        ╞════════════════╡
-        │ ["123", "45"]  │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ ["678", "910"] │
-        └────────────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_extract_all(pattern))
-
-    def count_match(self, pattern: str) -> Expr:
-        r"""
-        Count all successive non-overlapping regex matches.
-
-        Parameters
-        ----------
-        pattern
-            A valid regex pattern
-
-        Returns
-        -------
-        UInt32 array. Contain null if original value is null or regex capture nothing.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"foo": ["123 bla 45 asd", "xyz 678 910t"]})
-        >>> df.select(
-        ...     [
-        ...         pl.col("foo").str.count_match(r"\d").alias("count_digits"),
-        ...     ]
-        ... )
-        shape: (2, 1)
-        ┌──────────────┐
-        │ count_digits │
-        │ ---          │
-        │ u32          │
-        ╞══════════════╡
-        │ 5            │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ 6            │
-        └──────────────┘
-
-        """
-        return wrap_expr(self._pyexpr.count_match(pattern))
-
-    def split(self, by: str, inclusive: bool = False) -> Expr:
-        """
-        Split the string by a substring.
-
-        Parameters
-        ----------
-        by
-            Substring to split by.
-        inclusive
-            If True, include the split character/string in the results.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"s": ["foo bar", "foo-bar", "foo bar baz"]})
-        >>> df.select(pl.col("s").str.split(by=" "))
-        shape: (3, 1)
-        ┌───────────────────────┐
-        │ s                     │
-        │ ---                   │
-        │ list[str]             │
-        ╞═══════════════════════╡
-        │ ["foo", "bar"]        │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ ["foo-bar"]           │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ ["foo", "bar", "baz"] │
-        └───────────────────────┘
-
-        Returns
-        -------
-        List of Utf8 type
-
-        """
-        if inclusive:
-            return wrap_expr(self._pyexpr.str_split_inclusive(by))
-        return wrap_expr(self._pyexpr.str_split(by))
-
-    def split_exact(self, by: str, n: int, inclusive: bool = False) -> Expr:
-        """
-        Split the string by a substring into a struct of ``n`` fields.
-
-        If it cannot make ``n`` splits, the remaining field elements will be null.
-
-        Parameters
-        ----------
-        by
-            Substring to split by.
-        n
-            Number of splits to make.
-        inclusive
-            If True, include the split character/string in the results.
-
-        Examples
-        --------
-        >>> (
-        ...     pl.DataFrame({"x": ["a_1", None, "c", "d_4"]}).select(
-        ...         [
-        ...             pl.col("x").str.split_exact("_", 1).alias("fields"),
-        ...         ]
-        ...     )
-        ... )
-        shape: (4, 1)
-        ┌─────────────┐
-        │ fields      │
-        │ ---         │
-        │ struct[2]   │
-        ╞═════════════╡
-        │ {"a","1"}   │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ {null,null} │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ {"c",null}  │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ {"d","4"}   │
-        └─────────────┘
-
-
-        Split string values in column x in exactly 2 parts and assign
-        each part to a new column.
-
-        >>> pl.DataFrame({"x": ["a_1", None, "c", "d_4"]}).with_columns(
-        ...     [
-        ...         pl.col("x")
-        ...         .str.split_exact("_", 1)
-        ...         .struct.rename_fields(["first_part", "second_part"])
-        ...         .alias("fields"),
-        ...     ]
-        ... ).unnest("fields")
-        shape: (4, 3)
-        ┌──────┬────────────┬─────────────┐
-        │ x    ┆ first_part ┆ second_part │
-        │ ---  ┆ ---        ┆ ---         │
-        │ str  ┆ str        ┆ str         │
-        ╞══════╪════════════╪═════════════╡
-        │ a_1  ┆ a          ┆ 1           │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ null ┆ null       ┆ null        │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ c    ┆ c          ┆ null        │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┤
-        │ d_4  ┆ d          ┆ 4           │
-        └──────┴────────────┴─────────────┘
-
-        Returns
-        -------
-        Struct of Utf8 type
-
-        """
-        if inclusive:
-            return wrap_expr(self._pyexpr.str_split_exact_inclusive(by, n))
-        return wrap_expr(self._pyexpr.str_split_exact(by, n))
-
-    def replace(self, pattern: str, value: str, literal: bool = False) -> Expr:
-        r"""
-        Replace first matching regex/literal substring with a new string value.
-
-        Parameters
-        ----------
-        pattern
-            Regex pattern.
-        value
-            Replacement string.
-        literal
-             Treat pattern as a literal string.
-
-        See Also
-        --------
-        replace_all : Replace all matching regex/literal substrings.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"id": [1, 2], "text": ["123abc", "abc456"]})
-        >>> df.with_column(
-        ...     pl.col("text").str.replace(r"abc\b", "ABC")
-        ... )  # doctest: +IGNORE_RESULT
-        shape: (2, 2)
-        ┌─────┬────────┐
-        │ id  ┆ text   │
-        │ --- ┆ ---    │
-        │ i64 ┆ str    │
-        ╞═════╪════════╡
-        │ 1   ┆ 123ABC │
-        ├╌╌╌╌╌┼╌╌╌╌╌╌╌╌┤
-        │ 2   ┆ abc456 │
-        └─────┴────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_replace(pattern, value, literal))
-
-    def replace_all(self, pattern: str, value: str, literal: bool = False) -> Expr:
-        """
-        Replace all matching regex/literal substrings with a new string value.
-
-        Parameters
-        ----------
-        pattern
-            Regex pattern.
-        value
-            Replacement string.
-        literal
-             Treat pattern as a literal string.
-
-        See Also
-        --------
-        replace : Replace first matching regex/literal substring.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"id": [1, 2], "text": ["abcabc", "123a123"]})
-        >>> df.with_column(pl.col("text").str.replace_all("a", "-"))
-        shape: (2, 2)
-        ┌─────┬─────────┐
-        │ id  ┆ text    │
-        │ --- ┆ ---     │
-        │ i64 ┆ str     │
-        ╞═════╪═════════╡
-        │ 1   ┆ -bc-bc  │
-        ├╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌┤
-        │ 2   ┆ 123-123 │
-        └─────┴─────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_replace_all(pattern, value, literal))
-
-    def slice(self, start: int, length: int | None = None) -> Expr:
-        """
-        Create subslices of the string values of a Utf8 Series.
-
-        Parameters
-        ----------
-        start
-            Starting index of the slice (zero-indexed). Negative indexing
-            may be used.
-        length
-            Optional length of the slice. If None (default), the slice is taken to the
-            end of the string.
-
-        Returns
-        -------
-        Series of Utf8 type
-
-        Examples
-        --------
-        >>> df = pl.DataFrame({"s": ["pear", None, "papaya", "dragonfruit"]})
-        >>> df.with_column(
-        ...     pl.col("s").str.slice(-3).alias("s_sliced"),
-        ... )
-        shape: (4, 2)
-        ┌─────────────┬──────────┐
-        │ s           ┆ s_sliced │
-        │ ---         ┆ ---      │
-        │ str         ┆ str      │
-        ╞═════════════╪══════════╡
-        │ pear        ┆ ear      │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
-        │ null        ┆ null     │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
-        │ papaya      ┆ aya      │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
-        │ dragonfruit ┆ uit      │
-        └─────────────┴──────────┘
-
-        Using the optional `length` parameter
-
-        >>> df.with_column(
-        ...     pl.col("s").str.slice(4, length=3).alias("s_sliced"),
-        ... )
-        shape: (4, 2)
-        ┌─────────────┬──────────┐
-        │ s           ┆ s_sliced │
-        │ ---         ┆ ---      │
-        │ str         ┆ str      │
-        ╞═════════════╪══════════╡
-        │ pear        ┆          │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
-        │ null        ┆ null     │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
-        │ papaya      ┆ ya       │
-        ├╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
-        │ dragonfruit ┆ onf      │
-        └─────────────┴──────────┘
-
-        """
-        return wrap_expr(self._pyexpr.str_slice(start, length))
-
-
-class ExprDateTimeNameSpace:
-    """Namespace for datetime related expressions."""
-
-    def __init__(self, expr: Expr):
-        self._pyexpr = expr._pyexpr
-
-    def truncate(
-        self,
-        every: str | timedelta,
-        offset: str | timedelta | None = None,
-    ) -> Expr:
-        """
-        Divide the date/ datetime range into buckets.
-
-        The `every` and `offset` arguments are created with
-        the following string language:
-
-        1ns # 1 nanosecond
-        1us # 1 microsecond
-        1ms # 1 millisecond
-        1s  # 1 second
-        1m  # 1 minute
-        1h  # 1 hour
-        1d  # 1 day
-        1w  # 1 week
-        1mo # 1 calendar month
-        1y  # 1 calendar year
-
-        3d12h4m25s # 3 days, 12 hours, 4 minutes, and 25 seconds
-
-        Parameters
-        ----------
-        every
-            Every interval start and period length
-        offset
-            Offset the window
-
-        Returns
-        -------
-        Date/Datetime series
-
-        Examples
-        --------
-        >>> from datetime import timedelta, datetime
-        >>> start = datetime(2001, 1, 1)
-        >>> stop = datetime(2001, 1, 2)
-        >>> s = pl.date_range(start, stop, timedelta(minutes=30), name="dates")
-        >>> s
-        shape: (49,)
-        Series: 'dates' [datetime[ns]]
-        [
-            2001-01-01 00:00:00
-            2001-01-01 00:30:00
-            2001-01-01 01:00:00
-            2001-01-01 01:30:00
-            2001-01-01 02:00:00
-            2001-01-01 02:30:00
-            2001-01-01 03:00:00
-            2001-01-01 03:30:00
-            2001-01-01 04:00:00
-            2001-01-01 04:30:00
-            2001-01-01 05:00:00
-            2001-01-01 05:30:00
-            ...
-            2001-01-01 18:30:00
-            2001-01-01 19:00:00
-            2001-01-01 19:30:00
-            2001-01-01 20:00:00
-            2001-01-01 20:30:00
-            2001-01-01 21:00:00
-            2001-01-01 21:30:00
-            2001-01-01 22:00:00
-            2001-01-01 22:30:00
-            2001-01-01 23:00:00
-            2001-01-01 23:30:00
-            2001-01-02 00:00:00
-        ]
-        >>> s.dt.truncate("1h")
-        shape: (49,)
-        Series: 'dates' [datetime[ns]]
-        [
-            2001-01-01 00:00:00
-            2001-01-01 00:00:00
-            2001-01-01 01:00:00
-            2001-01-01 01:00:00
-            2001-01-01 02:00:00
-            2001-01-01 02:00:00
-            2001-01-01 03:00:00
-            2001-01-01 03:00:00
-            2001-01-01 04:00:00
-            2001-01-01 04:00:00
-            2001-01-01 05:00:00
-            2001-01-01 05:00:00
-            ...
-            2001-01-01 18:00:00
-            2001-01-01 19:00:00
-            2001-01-01 19:00:00
-            2001-01-01 20:00:00
-            2001-01-01 20:00:00
-            2001-01-01 21:00:00
-            2001-01-01 21:00:00
-            2001-01-01 22:00:00
-            2001-01-01 22:00:00
-            2001-01-01 23:00:00
-            2001-01-01 23:00:00
-            2001-01-02 00:00:00
-        ]
-        >>> assert s.dt.truncate("1h") == s.dt.truncate(timedelta(hours=1))
-
-        """
-        if offset is None:
-            offset = "0ns"
-        if isinstance(every, timedelta):
-            every = _timedelta_to_pl_duration(every)
-        if isinstance(offset, timedelta):
-            offset = _timedelta_to_pl_duration(offset)
-        return wrap_expr(self._pyexpr.date_truncate(every, offset))
-
-    def strftime(self, fmt: str) -> Expr:
-        """
-        Format Date/datetime with a formatting rule.
-
-        See `chrono strftime/strptime
-        <https://docs.rs/chrono/latest/chrono/format/strftime/index.html>`_.
-
-        """
-        return wrap_expr(self._pyexpr.strftime(fmt))
-
-    def year(self) -> Expr:
-        """
-        Extract year from underlying Date representation.
-
-        Can be performed on Date and Datetime.
-
-        Returns the year number in the calendar date.
-
-        Returns
-        -------
-        Year as Int32
-
-        """
-        return wrap_expr(self._pyexpr.year())
-
-    def quarter(self) -> Expr:
-        """
-        Extract quarter from underlying Date representation.
-
-        Can be performed on Date and Datetime.
-
-        Returns the quarter ranging from 1 to 4.
-
-        Returns
-        -------
-        Quarter as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.quarter())
-
-    def month(self) -> Expr:
-        """
-        Extract month from underlying Date representation.
-
-        Can be performed on Date and Datetime.
-
-        Returns the month number starting from 1.
-        The return value ranges from 1 to 12.
-
-        Returns
-        -------
-        Month as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.month())
-
-    def week(self) -> Expr:
-        """
-        Extract the week from the underlying Date representation.
-
-        Can be performed on Date and Datetime
-
-        Returns the ISO week number starting from 1.
-        The return value ranges from 1 to 53. (The last week of year differs by years.)
-
-        Returns
-        -------
-        Week number as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.week())
-
-    def weekday(self) -> Expr:
-        """
-        Extract the week day from the underlying Date representation.
-
-        Can be performed on Date and Datetime.
-
-        Returns the weekday number where monday = 0 and sunday = 6
-
-        Returns
-        -------
-        Week day as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.weekday())
-
-    def day(self) -> Expr:
-        """
-        Extract day from underlying Date representation.
-
-        Can be performed on Date and Datetime.
-
-        Returns the day of month starting from 1.
-        The return value ranges from 1 to 31. (The last day of month differs by months.)
-
-        Returns
-        -------
-        Day as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.day())
-
-    def ordinal_day(self) -> Expr:
-        """
-        Extract ordinal day from underlying Date representation.
-
-        Can be performed on Date and Datetime.
-
-        Returns the day of year starting from 1.
-        The return value ranges from 1 to 366. (The last day of year differs by years.)
-
-        Returns
-        -------
-        Day as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.ordinal_day())
-
-    def hour(self) -> Expr:
-        """
-        Extract hour from underlying DateTime representation.
-
-        Can be performed on Datetime.
-
-        Returns the hour number from 0 to 23.
-
-        Returns
-        -------
-        Hour as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.hour())
-
-    def minute(self) -> Expr:
-        """
-        Extract minutes from underlying DateTime representation.
-
-        Can be performed on Datetime.
-
-        Returns the minute number from 0 to 59.
-
-        Returns
-        -------
-        Minute as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.minute())
-
-    def second(self) -> Expr:
-        """
-        Extract seconds from underlying DateTime representation.
-
-        Can be performed on Datetime.
-
-        Returns the second number from 0 to 59.
-
-        Returns
-        -------
-        Second as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.second())
-
-    def nanosecond(self) -> Expr:
-        """
-        Extract seconds from underlying DateTime representation.
-
-        Can be performed on Datetime.
-
-        Returns the number of nanoseconds since the whole non-leap second.
-        The range from 1,000,000,000 to 1,999,999,999 represents the leap second.
-
-        Returns
-        -------
-        Nanosecond as UInt32
-
-        """
-        return wrap_expr(self._pyexpr.nanosecond())
-
-    def epoch(self, tu: EpochTimeUnit = "us") -> Expr:
-        """
-        Get the time passed since the Unix EPOCH in the give time unit
-
-        Parameters
-        ----------
-        tu : {'us', 'ns', 'ms', 's', 'd'}
-            Time unit.
-
-        """
-        if tu in DTYPE_TEMPORAL_UNITS:
-            return self.timestamp(tu)  # type: ignore[arg-type]
-        elif tu == "s":
-            return wrap_expr(self._pyexpr.dt_epoch_seconds())
-        elif tu == "d":
-            return wrap_expr(self._pyexpr).cast(Date).cast(Int32)
-        else:
-            raise ValueError(
-                f"tu must be one of {{'ns', 'us', 'ms', 's', 'd'}}, got {tu}"
-            )
-
-    def timestamp(self, tu: TimeUnit = "us") -> Expr:
-        """
-        Return a timestamp in the given time unit.
-
-        Parameters
-        ----------
-        tu : {'us', 'ns', 'ms'}
-            Time unit.
-
-        """
-        return wrap_expr(self._pyexpr.timestamp(tu))
-
-    def with_time_unit(self, tu: TimeUnit) -> Expr:
-        """
-        Set time unit of a Series of dtype Datetime or Duration.
-
-        This does not modify underlying data, and should be used to fix an incorrect
-        time unit.
-
-        Parameters
-        ----------
-        tu : {'ns', 'us', 'ms'}
-            Time unit for the ``Datetime`` Series.
-
-        """
-        return wrap_expr(self._pyexpr.dt_with_time_unit(tu))
-
-    def cast_time_unit(self, tu: TimeUnit) -> Expr:
-        """
-        Cast the underlying data to another time unit. This may lose precision.
-
-        Parameters
-        ----------
-        tu : {'ns', 'us', 'ms'}
-            Time unit for the ``Datetime`` Series.
-
-        """
-        return wrap_expr(self._pyexpr.dt_cast_time_unit(tu))
-
-    def with_time_zone(self, tz: str | None) -> Expr:
-        """
-        Set time zone for a Series of type Datetime.
-
-        Parameters
-        ----------
-        tz
-            Time zone for the `Datetime` Series.
-
-        """
-        return wrap_expr(self._pyexpr).map(
-            lambda s: s.dt.with_time_zone(tz), return_dtype=Datetime
-        )
-
-    def days(self) -> Expr:
-        """
-        Extract the days from a Duration type.
-
-        Returns
-        -------
-        A series of dtype Int64
-
-        """
-        return wrap_expr(self._pyexpr.duration_days())
-
-    def hours(self) -> Expr:
-        """
-        Extract the hours from a Duration type.
-
-        Returns
-        -------
-        A series of dtype Int64
-
-        """
-        return wrap_expr(self._pyexpr.duration_hours())
-
-    def minutes(self) -> Expr:
-        """
-        Extract the minutes from a Duration type.
-
-        Returns
-        -------
-        A series of dtype Int64
-
-        """
-        return wrap_expr(self._pyexpr.duration_minutes())
-
-    def seconds(self) -> Expr:
-        """
-        Extract the seconds from a Duration type.
-
-        Returns
-        -------
-        A series of dtype Int64
-
-        """
-        return wrap_expr(self._pyexpr.duration_seconds())
-
-    def milliseconds(self) -> Expr:
-        """
-        Extract the milliseconds from a Duration type.
-
-        Returns
-        -------
-        A series of dtype Int64
-
-        """
-        return wrap_expr(self._pyexpr.duration_milliseconds())
-
-    def nanoseconds(self) -> Expr:
-        """
-        Extract the nanoseconds from a Duration type.
-
-        Returns
-        -------
-        A series of dtype Int64
-
-        """
-        return wrap_expr(self._pyexpr.duration_nanoseconds())
-
-    def offset_by(self, by: str) -> Expr:
-        """
-        Offset this date by a relative time offset.
-
-        This differs from ``pl.col("foo") + timedelta`` in that it can
-        take months and leap years into account. Note that only a single minus
-        sign is allowed in the ``by`` string, as the first character.
-
-        Parameters
-        ----------
-        by
-            The offset is dictated by the following string language:
-
-            - 1ns   (1 nanosecond)
-            - 1us   (1 microsecond)
-            - 1ms   (1 millisecond)
-            - 1s    (1 second)
-            - 1m    (1 minute)
-            - 1h    (1 hour)
-            - 1d    (1 day)
-            - 1w    (1 week)
-            - 1mo   (1 calendar month)
-            - 1y    (1 calendar year)
-            - 1i    (1 index count)
-
-        Returns
-        -------
-        Date/Datetime expression
-
-        """
-        return wrap_expr(self._pyexpr.dt_offset_by(by))
-
-
 def expr_to_lit_or_expr(
     expr: (
         Expr
@@ -7122,56 +5123,6 @@ def expr_to_lit_or_expr(
             f"did not expect value {expr} of type {type(expr)}, maybe disambiguate with"
             " pl.lit or pl.col"
         )
-
-
-class ExprCatNameSpace:
-    """Namespace for categorical related expressions."""
-
-    def __init__(self, expr: Expr):
-        self._pyexpr = expr._pyexpr
-
-    def set_ordering(self, ordering: CategoricalOrdering) -> Expr:
-        """
-        Determine how this categorical series should be sorted.
-
-        Parameters
-        ----------
-        ordering : {'physical', 'lexical'}
-            Ordering type:
-
-            - 'physical' -> Use the physical representation of the categories to
-                determine the order (default).
-            - 'lexical' -> Use the string values to determine the ordering.
-
-        Examples
-        --------
-        >>> df = pl.DataFrame(
-        ...     {"cats": ["z", "z", "k", "a", "b"], "vals": [3, 1, 2, 2, 3]}
-        ... ).with_columns(
-        ...     [
-        ...         pl.col("cats").cast(pl.Categorical).cat.set_ordering("lexical"),
-        ...     ]
-        ... )
-        >>> df.sort(["cats", "vals"])
-        shape: (5, 2)
-        ┌──────┬──────┐
-        │ cats ┆ vals │
-        │ ---  ┆ ---  │
-        │ cat  ┆ i64  │
-        ╞══════╪══════╡
-        │ a    ┆ 2    │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌┤
-        │ b    ┆ 3    │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌┤
-        │ k    ┆ 2    │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌┤
-        │ z    ┆ 1    │
-        ├╌╌╌╌╌╌┼╌╌╌╌╌╌┤
-        │ z    ┆ 3    │
-        └──────┴──────┘
-
-        """
-        return wrap_expr(self._pyexpr.cat_set_ordering(ordering))
 
 
 def _prepare_alpha(
